+++ tests/testthat/test-get.R
@@ -7,21 +7,21 @@ dir_from <- test_path()
 test_that("conf_get() works (branch)", {
   # skip_on_travis()
   res <- confx::conf_get("col_names", dir_from = dir_from)
-  expect_is(res, "list")
+  expect_type(res, "list")
   expect_true(length(res) > 0)
 })
 
 test_that("conf_get() works (leaf)", {
   # skip_on_travis()
   res <- confx::conf_get("col_names/.col_id", dir_from = dir_from)
-  expect_is(res, "character")
+  expect_type(res, "character")
   expect_true(length(res) == 1)
 })
 
 test_that("conf_get() works (all)", {
   # skip_on_travis()
   res <- conf_get(dir_from = dir_from)
-  expect_is(res, "list")
+  expect_type(res, "list")
   expect_true(length(res) >= 1)
 })
 
@@ -34,14 +34,14 @@ test_that("conf_get() works (force from file)", {
   # skip_on_travis()
   warning("TODO-2019115-1: `force_from_file` hard to test, think about including an automatic load count or something similar")
   res <- conf_get(dir_from = dir_from, force_from_file = TRUE)
-  expect_is(res, "list")
+  expect_type(res, "list")
   expect_true(length(res) >= 1)
 })
 
 test_that("R_CONFIG_DIR", {
   Sys.setenv(R_CONFIG_DIR = file.path(dir_from, "subdir"))
   res <- conf_get(force_from_file = TRUE)
-  expect_is(res, "list")
+  expect_type(res, "list")
   expect_identical(res$test, "hello world")
 })
 
@@ -117,3 +117,17 @@ test_that("Temporary switch of config environment", {
   expect_identical(res, target)
 })
 
+# Absolute file paths -----------------------------------------------------
+
+test_that("Absolute file paths", {
+  "inst/configs/config.yml" %>% fs::path_abs() %>% print()
+  result <- conf_get(
+    "host/server_001",
+    from = "inst/configs/config.yml" %>% fs::path_abs(),
+    force_from_file = TRUE
+  )
+
+  expectation <- list(url = "https://dev-server-001.com", port = 8000L)
+
+  expect_identical(result, expectation)
+})